add api

+++ account/lib/app/modules/table_analyse/table_analyse_view.dart
@@ -27,9 +27,7 @@ class TableAnalysePage extends StatefulWidget {
 
 class _TableAnalysePageState extends State<TableAnalysePage> {
   final logic = Get.find<TableAnalyseLogic>();
-  final state = Get
-      .find<TableAnalyseLogic>()
-      .state;
+  final state = Get.find<TableAnalyseLogic>().state;
   final GlobalKey<SfDataGridState> _key = GlobalKey<SfDataGridState>();
 
   @override
@@ -69,7 +67,7 @@ class _TableAnalysePageState extends State<TableAnalysePage> {
                   changeTime: (start_, end_, isMonth_) {
                     logic.clear();
                     setState(
-                          () {
+                      () {
                         date = "$start_-01";
                       },
                     );
@@ -202,17 +200,16 @@ class DataSource extends DataGridSource {
   DataSource({required List<ConsumeData> data}) {
     _data = data
         .map<DataGridRow>(
-          (e) =>
-          DataGridRow(
+          (e) => DataGridRow(
             cells: [
               DataGridCell<String>(columnName: "时间", value: e.consumeDate),
-              DataGridCell<double>(columnName: "金额", value: e.amount),
-              DataGridCell<String>(
-                  columnName: "物品", value: e.consumptionName),
+              DataGridCell<double>(
+                  columnName: "金额", value: e.amount.toDouble()),
+              DataGridCell<String>(columnName: "物品", value: e.consumptionName),
               DataGridCell<String>(columnName: "地点", value: e.store),
             ],
           ),
-    )
+        )
         .toList();
   }
 
@@ -223,13 +220,12 @@ class DataSource extends DataGridSource {
   DataGridRowAdapter? buildRow(DataGridRow row) {
     return DataGridRowAdapter(
         cells: row.getCells().map<Widget>((dataGridCell) {
-          return Container(
-            padding: const EdgeInsets.all(16.0),
-            child: Text(dataGridCell.value.toString(),
-                maxLines: 1,
-                style: TextStyle(
-                    fontSize: 12.sp, color: const Color(0xff3D3D3D))),
-          );
-        }).toList());
+      return Container(
+        padding: const EdgeInsets.all(16.0),
+        child: Text(dataGridCell.value.toString(),
+            maxLines: 1,
+            style: TextStyle(fontSize: 12.sp, color: const Color(0xff3D3D3D))),
+      );
+    }).toList());
   }
 }
